Initial version of the documentation
There is still a few things to finish but i am going to push it so you
can see what i have been doing.
All the containers and files are getting setup. Documentation and
Utilities are being built.
Still need to integrate with the things duarte already did but ill do
that soon