Removing dbproj/ from the endpoint URL and adding clean.sh script
The clean.sh script its a utity to help during the development fase of
the project. It removes all of the images containers and files created
by the compose.sh script since we will need to rebuild the images for
testing purposes during the development.

Also removed the dbproj/ name from all the enpoints since i think its a
little too much verbose and will have little or no impact in the way
that the API will be used making it also more clear.